[contributing] Some notes about review / merge
For a while I've been wanting to make clear that the development here
should not turn into something too bureaucratic and we should not be
afraid at all of committing stuff to master. I guess just before the
move to an organization this update should happen.

I dunno about the wording tho, suggestions welcome.